EdIE-N/requirements.txt
matplotlib==3.0.2
numpy==1.15.1
PyYAML==5.3
scikit-learn==0.20.2
pytorch_transformers==1.2.0
tqdm==4.32.0
torch==1.3.0
mlconf==0.0.7
